revert https://github.com/dmlc/gluon-cv/pull/1249
